requirements.txt
astroid==1.6.5
cairocffi==0.8.1
CairoSVG==2.1.3
cffi==1.11.5
cssselect2==0.2.1
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0.6
django-import-export==1.0.1
et-xmlfile==1.0.1
gunicorn==19.8.1
html5lib==1.0.1
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
odfpy==1.3.6
openpyxl==2.5.4
pdfrw==0.4
Pillow==5.1.0
pkg-resources==0.0.0
pycparser==2.18
pylint==1.9.2
Pyphen==0.9.4
pytz==2018.4
PyYAML==3.12
six==1.11.0
tablib==0.12.1
tinycss2==0.6.1
unicodecsv==0.14.1
WeasyPrint==0.42.3
webencodings==0.5.1
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
